Make 3d map rendering better

+++ utils/Map.h
@@ -225,8 +225,8 @@ public:
         inline bool IsTraversable(long x, long y)const{ return x>=0&&x<width&&y>=0&&y<height&&(GetTerrainType(x,y)&0xF0); }
 	
 	void OpenGLDraw(tDisplay how = kPolygons) const;
-	bool GetOpenGLCoord(int _x, int _y, GLdouble &x, GLdouble &y, GLdouble &z, GLdouble &radius) const;
-        bool GetOpenGLCoord(int _x, int _y, int _z, GLdouble &x, GLdouble &y, GLdouble &z, GLdouble &radius) const{return  GetOpenGLCoord(_x,_y,x,y,z,radius);}
+	bool GetOpenGLCoord(float _x, float _y, GLdouble &x, GLdouble &y, GLdouble &z, GLdouble &radius) const;
+        bool GetOpenGLCoord(float _x, float _y, float _z, GLdouble &x, GLdouble &y, GLdouble &z, GLdouble &radius) const{return  GetOpenGLCoord(_x,_y,x,y,z,radius);}
 	void GetPointFromCoordinate(point3d loc, int &px, int &py) const;
 	double GetCoordinateScale();
         bool LineOfSight(int x, int y, int _x, int _y) const;
